+++ CHANGELOG.md
@@ -4,11 +4,11 @@
 
 # Unreleased
 
-No current changes
+- Respect CLI context order
 
 # v4.9.2
 
-- Always use first CLI argument as target to allow non-component targets
+- Determine CLI target based on order to allow non-component targets
 - Ensure that config field is always reloaded from index to avoid incorrect recomputation
 
 # v4.9.1